refactor: BatchScheduler try-catch 중복 제거 및 DTO 패키지 정리
기술부채 해결:
- BatchScheduler: 불필요한 중첩 try-catch 제거
- CheckedLogicExecutor.executeUncheckedVoid + 메서드 참조로 단순화
- 코드 16줄 감소 (109→93줄)

DTO 패키지 정리:
- auth/ 하위 디렉토리 제거, web/dto/ 직접 배치
- Java Record 호환성 메서드 추가
