+++ module-web/src/main/kotlin/maple/expectation/web/dto/LoginRequest.kt
@@ -22,6 +22,13 @@ data class LoginRequest(
     @field:NotBlank(message = "캐릭터 닉네임은 필수입니다.")
     val userIgn: String
 ) {
+    // Java Record 호환 메서드 (기존 Java 코드와의 호환성 유지)
+    /** @return API Key (Java record-style accessor) */
+    fun apiKey(): String = apiKey
+
+    /** @return 사용자 캐릭터 닉네임 (Java record-style accessor) */
+    fun userIgn(): String = userIgn
+
     /**
      * API Key 마스킹된 문자열 반환 (로그 보안)
      *
